requirements.txt
requests
uuid
jsonlines
typing
datetime==5.2
huggingface_hub==0.16.4
datasets==2.13.1
evaluate==0.4.0
pytorchvideo==0.1.5
rouge-score==0.1.2
tokenizers==0.13.3
torchdata==0.6.1
torchinfo==1.8.0
torchtext==0.15.2
transformers==4.30.2
tqdm==4.65.0
nltk==3.8.1
pandas==2.0.3
